+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Minfraud Changelog
+# Changelog
 
 ## v2.0.0